pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "persona"
version = "0.1.0"
description = "Persona-based survey system with WebSocket support"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "datasets>=2.16.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.0.5",
    "openai>=1.84.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-socketio>=5.11.0",
    "fastapi-socketio>=0.0.10",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/persona"]
